color: Re-factor Wraase modes
`ColorSSTV` parent class actually implements the scan-line encoding we
need, however the challenge is that it seems the sync pulse requirements
differ for SC2-120 and SC2-180 just slightly.

I haven't figured out why, partially because there seems to be little in
the way of clear (and correct!) docs as to how SC2-120 is supposed to work.